emotion/requirements.txt
# HSEmotion 측 요구사항에 맞춰 timm 버전을 고정하는 것을 권장
# (원 스니펫 주석: pip install --upgrade "timm==0.8.19.dev0")

torch>=2.0.0
facenet-pytorch>=2.6.0
hsemotion>=0.3.0
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0
timm==0.8.19.dev0
Pillow>=10.0.0
transformers>=4.44.0
accelerate>=0.26.0
sentencepiece>=0.1.99
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
sentence-transformers>=2.6.0
requests>=2.31.0
python-dotenv>=1.0.0
tf-keras

# Pose 분석 API까지 사용하려면 다음 추가 의존성이 필요합니다.
mediapipe>=0.10
ultralytics>=8.0
deepface>=0.0.75
